Dictionary Tests
Used both an O(n) and O(1) method from the book. The algorithm can easily be used with each dictionary action by replacing a single line. I am getting tiny numbers in return for the time so my computer processes are actually messing with them. Will have to check on the computers in the lab.